+++ docs/requirements.txt
@@ -17,3 +17,4 @@ tensorboard
 attridict
 fedgraph
 dtaidistance
+gdown
