+++ src/server_class.py
@@ -8,7 +8,6 @@ import torch
 from dtaidistance import dtw
 
 from src.gnn_models import GCN, GNN_LP, AggreGCN, GCN_arxiv, SAGE_products
-from src.trainer_class import Trainer_General
 
 
 class Server:
